software/pyproject.toml
[project]
name = "cr-bio-software"
version = "0.1.0"
description = "Course management utilities for Biology courses at College of the Redwoods"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "markdown>=3.5.0",
    "weasyprint>=60.0",
    "gtts>=2.5.0",
    "requests>=2.31.0",
    "speechrecognition>=3.10.0",
    "pydub>=0.25.1",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "markdown2>=2.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "mypy>=1.7.0",
    "ruff>=0.1.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "-v",
    "-W",
    "ignore::DeprecationWarning:speech_recognition",
]
filterwarnings = [
    "ignore::DeprecationWarning:speech_recognition",
]
markers = [
    "requires_internet: marks tests that require internet connection (deselect with '-m \"not requires_internet\"')",
    "requires_api: marks tests that require external API access",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py311"
